+++ package.json
@@ -64,6 +64,7 @@
     "@typescript-eslint/eslint-plugin": "5.30.0",
     "@typescript-eslint/parser": "5.30.0",
     "eslint": "8.18.0",
+    "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-prettier": "4.1.0",
     "jest": "28.1.1",
     "prettier": "2.7.1",
