+++ src/accounts/accounts.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common'
+import { AccountsController } from './accounts.controller'
+
+@Module({
+  imports: [],
+  controllers: [AccountsController],
+})
+export class AccountsModule {}
