+++ src/app.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common'
 import { TypeOrmModule } from '@nestjs/typeorm'
 import { ConfigModule } from '@nestjs/config'
-import { BrokerModule } from './broker/broker.module'
+import { AccountsModule } from './accounts/accounts.module'
 import { AlpacaModule } from './alpaca/alpaca.module'
 import typeOrmConfig from './typeorm.config'
 import alpacaModuleConfig from './alpaca.config'
@@ -11,7 +11,7 @@ import alpacaModuleConfig from './alpaca.config'
     ConfigModule.forRoot(),
     TypeOrmModule.forRoot(typeOrmConfig()),
     AlpacaModule.register(alpacaModuleConfig()),
-    BrokerModule,
+    AccountsModule,
   ],
   providers: [],
 })
